feat: mannual level generation

+++ src/entity/derived/Cactus.java
@@ -13,7 +13,7 @@ public class Cactus extends Box {
      * @param x    the x
      * @param y    the y
      */
-    public Cactus(String name, int x, int y) {
+    public Cactus(String name, double x, double y) {
         super(name, x, y);
         this.setFitHeight(50);
         this.setFitWidth(50);
